move focusRadioH3 and update imports with new location
