+++ src/applications/appeals/996/utils/focus.js
@@ -1,22 +1,4 @@
-import {
-  focusElement,
-  scrollTo,
-  scrollToTop,
-  defaultFocusSelector,
-  waitForRenderThenFocus,
-} from 'platform/utilities/ui';
-import { $ } from 'platform/forms-system/src/js/utilities/ui';
-
-export const focusRadioH3 = () => {
-  scrollToTop();
-  const radio = $('va-radio');
-  if (radio) {
-    // va-radio content doesn't immediately render
-    waitForRenderThenFocus('h3', radio.shadowRoot);
-  } else {
-    focusElement(defaultFocusSelector);
-  }
-};
+import { scrollTo, waitForRenderThenFocus } from 'platform/utilities/ui';
 
 export const focusAlertH3 = () => {
   scrollTo('h3');
